+++ src/ws_messages/outgoing.rs
@@ -1,3 +1,5 @@
+use jiff::Zoned;
+// use jiff::Zoned;
 use serde::{Deserialize, Serialize};
 use tokio_tungstenite::tungstenite::Message;
 
@@ -12,6 +14,7 @@ pub struct PiStatus {
     pub screen_status: Option<ScreenStatus>,
     pub time_off: (i8, i8),
     pub time_on: (i8, i8),
+    pub timezone: String,
     pub uptime_app: u64,
     pub uptime_ws: u64,
     pub uptime: usize,
@@ -20,11 +23,13 @@ pub struct PiStatus {
 /// Combined pi into and current set alarms
 impl PiStatus {
     pub fn new(sysinfo: SysInfo, uptime_ws: u64) -> Self {
+        let zone = Zoned::now();
         Self {
             ip_address: sysinfo.ip_address,
             screen_status: sysinfo.screen_status,
             time_off: sysinfo.time_off,
             time_on: sysinfo.time_on,
+            timezone: zone.time_zone().iana_name().unwrap_or("Etc/UTC").to_owned(),
             uptime_app: sysinfo.uptime_app,
             uptime: sysinfo.uptime,
             uptime_ws,