+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# <a href='https://github.com/mrjackwills/screen_control_backend/releases/tag/v0.1.2'>v0.1.2</a>
+### 2025-04-07
+
+### Chores
++ dependencies updated, [40177612](https://github.com/mrjackwills/screen_control_backend/commit/40177612fc0a6659887646cff5475410ce12780c)
++ Rust 1.86.0 linting, [7412a3b5](https://github.com/mrjackwills/screen_control_backend/commit/7412a3b5f86bfe09d4a4e2f89a54fcd520a5b899)
+
+### Features
++ send timezone to client, [5f56ee01](https://github.com/mrjackwills/screen_control_backend/commit/5f56ee013de4cf1137f3ae5f90c37610f4696b27)
+
 # <a href='https://github.com/mrjackwills/screen_control_backend/releases/tag/v0.1.1'>v0.1.1</a>
 ### 2025-02-26